% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wl.R
\name{color.bar}
\alias{color.bar}
\title{Make a color spectrum graphic}
\usage{
color.bar(
  bc,
  min,
  max = -min,
  nticks = 5,
  ticks = seq(min, max, len = nticks),
  title = ""
)
}
\arguments{
\item{bc}{barColors, the character variable of colors across the spectrum}

\item{min}{numeric; the value corresponding to the first color in bc (usually -1.5)}

\item{max}{numeric; the value corresponding to the last color in bc (usually 1.5)}

\item{nticks}{default 5; number of ticks}

\item{ticks}{variable of where you want the ticks to show; default is evenly spaced}

\item{title}{character to set title}
}
\value{
A color bar for use in making graphics
}
\description{
Create color bars (taken from http://www.colbyimaging.com/wiki/statistics/color-bars)
}
