% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_plot.R
\name{error_plot_gradient}
\alias{error_plot_gradient}
\title{Error Plot for Univariate Comparisons with color gradient}
\usage{
error_plot_gradient(
  x,
  y,
  color = y,
  xlab = "",
  ylab = "",
  color.str = "",
  title = ylab,
  axis.text.x.size = 20,
  axis.text.y.size = 20,
  axis.text.x.angle = 90,
  title.size = 20,
  text.size = 20,
  low = "blue",
  mid = "white",
  high = "red3"
)
}
\arguments{
\item{x}{string or factor variable of your experimental groups by sample}

\item{y}{numeric variable for y axis}

\item{color}{defaults to y; numeric to color your data points by gradient}

\item{xlab, ylab}{default ""; axis labels}

\item{color.str}{default ""; a string to label the color legend}

\item{title}{default ylab; sets title of the scores plot}
}
\value{
A ggplot2 object of the error plot
}
\description{
display univariate mean/SEM comparisons across groups
}
