% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wl.R
\name{heatmap_wl}
\alias{heatmap_wl}
\title{Heatmap3 made simple(r)}
\usage{
heatmap_wl(
  data,
  mar = c(5, 2),
  title = "",
  xlab = "",
  ylab = "",
  color1 = "blue",
  color2 = "white",
  color3 = "red2",
  cex_r = 1,
  cex_c = 1,
  z_lim = 1.5,
  clust_r = FALSE,
  clust_c = FALSE,
  labRow = rownames(data),
  labCol = colnames(data)
)
}
\arguments{
\item{data}{data frame or matrix for heatmap; will be scaled across columns}

\item{mar}{default c(5,2); margin for column and row names}

\item{title}{default ""; set a title}

\item{xlab, ylab}{default ""; set axis labels}

\item{color1, color2, color3}{default blue to red color scale; set to change heatmap colors}

\item{cex_r, cex_c}{default 1; numeric scale factor for axis text}

\item{z_lim}{default 1.5; numeric to set max/min colors in z-score scale}

\item{clust_r}{set to TRUE to cluster rows}

\item{clust_c}{set to TRUE to cluster columns}

\item{labRow}{default is rownames(data); sets row labels}

\item{labCol}{default is colnames(data); sets col labels}
}
\value{
A heatmap of your data
}
\description{
Create a heatmap with minimal input
}
