% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_chart.R
\name{loadings_chart}
\alias{loadings_chart}
\title{Loadings chart for multivariate analyses (PCA,PLS)}
\usage{
loadings_chart(loadings, component_str, lim = 1, cex = 1, col = "lightblue")
}
\arguments{
\item{loadings}{loadings variable (e.g. P1 or P2 output from rotate_opls)
Set rownames() for analyte labels}

\item{component_str}{string with name of component (e.g. "LV1" or "LV2")}

\item{lim}{default 1; axis limit}

\item{cex}{numeric scaling factor for labels}

\item{col}{default "lightblue"; set for bar colors in barplot()}
}
\value{
A ggplot object of the loadings chart in ascending order
}
\description{
Used to create loadings bar plots sorted by value and scaled by diving by max value.
}
