% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_chart.R
\name{loadings_chart2}
\alias{loadings_chart2}
\title{Alternate loadings chart for multivariate analyses (PCA,PLS) using ggpubr}
\usage{
loadings_chart2(loadings, component_str, x_size = 10)
}
\arguments{
\item{loadings}{loadings variable (e.g. P1 or P2 output from rotate_opls)
Set rownames() for analyte labels}

\item{component_str}{string with name of component (e.g. "LV1" or "LV2")}

\item{x_size}{default 10; text size for x axis text}
}
\value{
A ggplot object of the loadings chart in ascending order
}
\description{
Used to create loadings bar plots sorted by value and scaled by diving by max value.
}
