% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOCV.R
\name{opls_LOOCV}
\alias{opls_LOOCV}
\title{Iterative PLS / PLSR with opls() using leave-n out cross validation}
\usage{
opls_LOOCV(x, y, n = 1, runs = 100)
}
\arguments{
\item{x}{x input for opls(), the numeric matrix of samples (rows) and analytes (columns)}

\item{y}{y input for opls(), either an array of regression variable values (PLSR) or an array of factors for group membership (PLSDA)}

\item{n}{default 1; the number of randomized samples to be left out per iteration, must be an integer
Recommend 15\% of samples}

\item{runs}{default 100; 100, the number of iterations to conduct leave-n out cross validation, must be an integer}
}
\value{
A list with the following objects:
loadings = loadings in LV1 (first column) and LV2 (second column) from opls() of all samples (not LOOCV)
scores = scores in LV1 (first column) and LV2 (second column) from opls() of all samples (not LOOCV)
loadings_LOOCV = mean values of loadings in LV1 (first column) and LV2 (second column) from LOOCV iterations
LV1_mat = matrix of LV1 output of each LOOCV iteration (rows), for use in calculating standard deviation after rotation
LV2_mat = matrix of LV2 output of each LOOCV iteration (rows), for use in calculating standard deviation after rotation
}
\description{
Conducts partial least squares (PLS) regression (PLSR) or discriminant analysis (PLS-DA) using ropls::opls()
Iteratively computes PLS scores and loadings runs=100 times, each time leaving out n=1 samples
Outputs all iterations of scores and loadings for computing standard deviation

note: the term LOOCV (leave one out cross validation) is used broadly below, but changing value of n in input will leave-n out for all output
}
