% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_opls.R
\name{rotate_opls}
\alias{rotate_opls}
\title{Rotate an opls object}
\usage{
rotate_opls(plsOut, degrees = 0, swap = F, flip_y = F, flip_x = F)
}
\arguments{
\item{plsOut}{output object from opls() function, must contain only 2 components
Ex. plsOut = opls(data,predI=2) for Principal Component Analysis
plsOut = opls(data,y=pheno,predI=2) for PLS}

\item{degrees}{default 0; specified degrees of clockwise rotation}

\item{swap}{default FALSE; set to TRUE to swap x and y axes}

\item{flip_y}{default FALSE; set to TRUE to flip data across y axis}

\item{flip_x}{default FALSE; set to TRUE to flip data across x axis}
}
\value{
A list with the following objects:
T1 = scores in rotated LV1
T2 = scores in rotated LV2
P1 = loadings in rotated LV1
P2 = loadings in rotated LV2
rotmat = matrix of rotation by specified degrees
}
\description{
Used to rotate scores and loadings on opls objects by specified degrees, then perform axis swaps and/or flips. Output is list of scores, loadings, and rotation matrix.
}
