% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOCV.R
\name{rotate_opls_LOOCV}
\alias{rotate_opls_LOOCV}
\title{Rotate opls_LOOCV ouput}
\usage{
rotate_opls_LOOCV(
  opls_LOOCV_out,
  degrees = 0,
  swap = F,
  flip_y = F,
  flip_x = F
)
}
\arguments{
\item{opls_LOOCV_out}{object output from opls.LOOCV()}

\item{degrees}{default 0; specified degrees of clockwise rotation}

\item{swap}{default FALSE; set to TRUE to swap x and y axes}

\item{flip_y}{default FALSE; set to TRUE to flip data across y axis}

\item{flip_x}{default FALSE; set to TRUE to flip data across x axis}
}
\value{
A list with the following objects:
T1 = scores in rotated LV1
T2 = scores in rotated LV2
P1 = loadings in rotated LV1
P2 = loadings in rotated LV2
P1_LOOCV = mean value of loadings in LV1, calculated across LOOCV iterations
P2_LOOCV = mean value of loadings in LV2, calculated across LOOCV iterations
sd_LV1_LOOCV = standard deviation of each loading in LV1, calculated across LOOCV iterations
sd_LV2_LOOCV = standard deviation of each loading in LV2, calculated across LOOCV iterations
rotmat = matrix of rotation by specified degrees
}
\description{
Rotates output from function opls.LOOCV by specified degrees and performs specified axis swap/flips
Computes and returns standard deviations of loadings across all LOOCV iterations

Ex. opls_LOOCV_out = opls_LOOCV(data,predI=2)
rotated_opls = rotate_opls_LOOCV(opls_LOOCV_out,30) # 30 degree rotation
}
