% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_reg.R
\name{scatter_reg}
\alias{scatter_reg}
\title{Scatter plot with regression}
\usage{
scatter_reg(
  x,
  y,
  xlab = "",
  ylab = "",
  title = "",
  point_color = "black",
  point_size = 4,
  line_color = "red",
  line_size = 1.5,
  linetype = "longdash",
  axisfontsize = 14,
  method = "lm",
  showRSQ = TRUE,
  showP = TRUE,
  text_position = "lower right",
  xintercept = NA
)
}
\arguments{
\item{x}{numeric x variable}

\item{y}{numeric y variable}

\item{xlab, ylab}{default ""; set axis labels}

\item{title}{default ""; set a title}

\item{point_size}{default 4; numeric size of points}

\item{line_color}{default "red"; regression line color}

\item{line_size}{default 1.5; regression line size}

\item{linetype}{default "longdash"; regression line type}

\item{axisfontsize}{default 14; axis font size}

\item{method}{default is "lm"; set regression method}

\item{showRSQ, showP}{default is TRUE; set to FALSE to hide regression statistics}

\item{text_position}{default is "lower right"
Allowable values are "lower left", "upper right", "upper left"}

\item{xintercept}{default is NA; change to numeric to show vertical line at ideal loading}
}
\value{
A y vs x scatter plot with regression line and OLS regression statistics
}
\description{
Make scatter plot with a regression line and display statistics
Please note: R^2 and p-value do NOT change with method input - always OLS "lm" method
}
