% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_plot.R
\name{scores_plot}
\alias{scores_plot}
\title{Make a Scores Plot}
\usage{
scores_plot(
  T1,
  T2,
  color,
  color.str = "",
  analysis = "Scores Plot",
  plot.title = NULL
)
}
\arguments{
\item{T1, T2}{sample scores in your two components (e.g.: T1=PC1 and T2=PC2)}

\item{color}{string or factor variables of your experimental groups by sample}

\item{color.str}{default ""; a string to label the color legend}

\item{analysis}{default "Scores Plot"; updates title and axis labels; title can be superseded by plot.title
Allowed values are "PCA" to set axis labels to "PC" or "PLS" to set axis labels to "LV"}

\item{plot.title}{Set title of the scores plot; if not given, defaults to analysis}
}
\value{
A ggplot2 object of the scores plot
}
\description{
Used to create scores plot colored by your categorical variable of interest.
}
