% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_plot.R
\name{scores_plot_gradient}
\alias{scores_plot_gradient}
\title{Make a Scores Plot with Gradient Colored Points}
\usage{
scores_plot_gradient(
  T1,
  T2,
  color,
  color.str,
  analysis = "PLSDA",
  plot.title = NULL
)
}
\arguments{
\item{T1, T2}{Sample scores in your two components (e.g.: PC1 and PC2)}

\item{color}{A character or factor array of your experimental groups by sample}

\item{color.str}{Text to label the color legend}

\item{analysis}{Set to "PCA" to change axis labels to PC instead of LV; defaults to "PLSDA" with LV axes; sets default plot title}

\item{plot.title}{Set title of the scores plot; if not given, defaults to analysis}
}
\value{
A ggplot2 object of the scores plot
}
\description{
Used to create scores plot colored by quantitative variable of interest.
}
