% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_opls.R
\name{rotate_opls}
\alias{rotate_opls}
\title{Rotate an opls object}
\usage{
rotate_opls(plsOut, degrees = 0, swap = F, flip_y = F, flip_x = F)
}
\arguments{
\item{plsOut}{output object from opls() function, must contain only 2 components
Ex. plsOut = opls(data,predI=2) for Principal Component Analysis
plsOut = opls(data,y=pheno,predI=2) for PLS}

\item{degrees}{default 0; numeric specifying how many degrees to rotate clockwise}

\item{swap}{default FALSE; set to TRUE to swap x and y axes}

\item{flip_y}{default FALSE; set to TRUE to flip data across y axis}

\item{flip_x}{default FALSE; set to TRUE to flip data across x axis}
}
\value{
A list of rotated scores, loadings, and the rotation matrix used.
out$T1 and out$T2 call the rotated scores in the 1st and 2nd component, respectively
out$P1 and out$P2 call the rotated loadings in the 1st and 2nd component, respectively
out$rotmat calls the rotation matrix used to transform the original data
}
\description{
Used to rotate scores and loadings on opls objects by specified degrees, then perform axis swaps and/or flips. Output is list of scores, loadings, and rotation matrix.
}
