% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_plot.R
\name{scores_plot_ellipse}
\alias{scores_plot_ellipse}
\title{Scores Plot with Confidence Ellipses}
\usage{
scores_plot_ellipse(
  T1,
  T2,
  color,
  color.str = "",
  analysis = "Scores Plot",
  plot.title = NULL,
  level = 0.95
)
}
\arguments{
\item{T1, T2}{sample scores in your two components (e.g.: T1=PC1 and T2=PC2)}

\item{color}{string or factor variables of your experimental groups by sample}

\item{color.str}{default ""; a string to label the color legend}

\item{analysis}{default "Scores Plot"; updates title and axis labels; title can be superseded by plot.title
Allowed values are "PCA" to set axis labels to "PC" or "PLS" to set axis labels to "LV"}

\item{plot.title}{Set title of the scores plot; if not given, defaults to analysis}

\item{level}{confidence level of ellipse, default set to 0.95}
}
\value{
A ggplot2 object of the scores plot
}
\description{
Create scores plot colored by categorical variable of interest with ellipses showing confidence level of each group.
}
